CA-118 split output subdirectory
